+++ core/src/main/java/cn/edu/tsinghua/iginx/engine/logical/utils/LogicalFilterUtils.java
@@ -2,6 +2,7 @@ package cn.edu.tsinghua.iginx.engine.logical.utils;
 
 import cn.edu.tsinghua.iginx.engine.physical.memory.execute.utils.ExprUtils;
 import cn.edu.tsinghua.iginx.engine.shared.KeyRange;
+import cn.edu.tsinghua.iginx.engine.shared.expr.*;
 import cn.edu.tsinghua.iginx.engine.shared.operator.filter.*;
 import cn.edu.tsinghua.iginx.metadata.entity.ColumnsInterval;
 import cn.edu.tsinghua.iginx.metadata.entity.FragmentMeta;
@@ -870,5 +871,26 @@ public class LogicalFilterUtils {
           }
         });
     return paths;
+  }
+
+  /**
+   * 将Filter分解为一系列AND连接的子条件，如果顶层Filter不是AND连接的话，转换为CNF再分解
+   *
+   * @param filter 待分解的Filter
+   * @return 分解后的子条件列表
+   */
+  public static List<Filter> splitFilter(Filter filter) {
+    List<Filter> splitFilter = new ArrayList<>();
+    if (filter.getType() != FilterType.And) {
+      filter = toCNF(filter);
+    }
+
+    if (filter.getType() != FilterType.And) {
+      splitFilter.add(filter);
+    } else {
+      AndFilter andFilter = (AndFilter) filter;
+      splitFilter.addAll(andFilter.getChildren());
+    }
+    return splitFilter;
   }
 }
